fixed merge conflict

+++ app/models/user.py
@@ -17,7 +17,7 @@ class User(db.Model, UserMixin):
     shop_logo_img = db.Column(db.String(255))
     shop_splash_img = db.Column(db.String(255))
     category = db.Column(db.String(255))
-    phone_number = db.Column(db.Numeric(10), nullable=False, unique=True)
+    phone_number = db.Column(db.Integer, nullable=False, unique=True)
     zipcode = db.Column(db.String, nullable=False,)
     hashed_password = db.Column(db.String(255), nullable=False)
